+++ src/index.ts
@@ -1,58 +1,11 @@
-import { Hono } from "hono";
-import { zValidator } from "@hono/zod-validator";
-import { captureError } from "@cfworker/sentry";
-import z from "zod";
-import { useOctoApp } from "./octo";
+/**
+ * App entry point for Cloudflare.
+ *
+ * This looks stupid but that's because hono router naively implements cloudflare's
+ * fetch interface. We might add other runtime (probably nodejs) in the future
+ * if there's any need with initialization logic in this file.
+ */
 
-const app = new Hono<{ Bindings: Env }>();
-
-app.post(
-  "/github/webhook",
-  zValidator(
-    "header",
-    z.object({
-      "x-github-hook-id": z.string(),
-      "x-github-event": z.string(),
-      "x-hub-signature-256": z.string(),
-    }),
-  ),
-  async (c) => {
-    const octoApp = useOctoApp(c.env);
-
-    const headers = c.req.valid("header");
-
-    await octoApp.webhooks.verifyAndReceive({
-      id: headers["x-github-hook-id"],
-      name: headers["x-github-event"] as any, // FIXME: Verify enum on zod
-      signature: headers["x-hub-signature-256"],
-      payload: await c.req.text(),
-    });
-
-    return c.text("ok", 202);
-  },
-);
-
-app.notFound((c) => {
-  return c.text("Not found", 404);
-});
-
-app.onError((err, c) => {
-  console.error("Router raised exception", err);
-
-  if (c.env.SENTRY_DSN) {
-    const { posted } = captureError({
-      sentryDsn: c.env.SENTRY_DSN,
-      environment: "prod",
-      release: "release",
-      err,
-      request: c.req.raw,
-      user: "",
-    });
-
-    c.executionCtx.waitUntil(posted);
-  }
-
-  return c.text("Internal server error", { status: 500 });
-});
+import app from "./router";
 
 export default app;
