+++ src/utils.ts
@@ -1,3 +1,8 @@
+/**
+ * Assert given value is not nullish (null or undefined).
+ *
+ * Borrowed from rust for satisfying typescript checker.
+ */
 export function unwrap<T>(
   value: T,
   message: string = "Tried to unwrap a nullish value",
